Improve initialization process for NimBLE client
Rename setupAndStartBleScans to setupBleScans since it only
configures the BLE scanning.

Create new internal function to start BLE scans that gets called
from the begin() method.

keepAlive doesn't need to be called on init to start the scanning.
It can be used to re-start in case of errors as documented.