Improve BLE interface
Use uint64_t as interface between NimBLE client and SensiScan
as we only convert it into a invalid string and back into an
uint64_t.

Add constructor and destructure to NimBLE client to better
represent the life-cycle of this class.

Remove squash method from SensiScan as it is not required anymore.